Fix Bugs (#73)
* fix mixlora slicing on macOS

* add optimal approach of mixlora to chatglm

* update requirements

* remove deprecated expand side

+++ mlora/models/modeling_llama.py
@@ -352,8 +352,12 @@ class LlamaMLP(LLMFeedForward):
     def _mixlora_forward(
         self, moe_name, act_fn, expert_mask, hidden_states, input_dtype
     ):
-        common_w1 = self.w1_.base_layer_.forward(hidden_states.to(input_dtype))
-        common_w3 = self.w3_.base_layer_.forward(hidden_states.to(input_dtype))
+        common_w1 = self.w1_.base_layer_.forward(hidden_states.to(input_dtype)).to(
+            hidden_states.dtype
+        )
+        common_w3 = self.w3_.base_layer_.forward(hidden_states.to(input_dtype)).to(
+            hidden_states.dtype
+        )
         final_expert_states = []
         for expert_idx in range(expert_mask.shape[0]):
             _, top_x = torch.where(expert_mask[expert_idx])
